added routers for different functionalities
