+++ app/routers/user.py
@@ -1,0 +1,31 @@
+from fastapi import FastAPI, Depends, status, HTTPException, APIRouter
+from .. import models, schemas, utils
+from sqlalchemy.orm import Session
+from ..database import get_db
+
+router = APIRouter(
+    prefix="/users",
+    tags=['Users']
+)
+
+
+@router.post("/", status_code=status.HTTP_201_CREATED, response_model=schemas.UserOut)
+def create_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
+    # hashing the user password
+    hashed_password = utils.hash(user.password)
+    user.password = hashed_password    
+    new_user = models.User(**user.dict())
+    db.add(new_user)
+    db.commit()
+    db.refresh(new_user)
+
+    return new_user
+
+
+@router.get("/{id}", response_model=schemas.UserOut)
+def get_user(id: int, db: Session = Depends(get_db)):
+    user = db.query(models.User).get(id)
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail = f"user with id: {id} does not exist!")
+
+    return user
